Add User Registration

+++ server/src/utils/cloudinary.js
@@ -1,27 +1,36 @@
-const {v2:cloudinary} = require("cloudinary");
+const { v2: cloudinary } = require("cloudinary");
 const fs = require("fs");
 
-cloudinary.config({
-    cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
-    api_key: process.env.CLOUDINARY_API_KEY,
-    api_secret: process.env.CLOUDINARY_API_SECRET
+
+cloudinary.config({ 
+  cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
+  api_key: process.env.CLOUDINARY_API_KEY, 
+  api_secret: process.env.CLOUDINARY_API_SECRET 
 });
 
-const uploadOnCloudinary = async (localFilePath)=>{
+const uploadOnCloudinary = async (localFilePath) => {
     try {
-        if(!localFilePath) return null;
-        //upload on cloudinary
+        if (!localFilePath) return null;
+        //upload the file on cloudinary
+
+        console.log("Uploading:", localFilePath);
+        console.log("Type:", typeof localFilePath);
+
         const response = await cloudinary.uploader.upload(localFilePath, {
             resource_type: "auto"
-        });
-        //file has been uploaded successfully
-        console.log("File is Uploaded on Cloudinary... ", response.url);
-        fs.unlinkSync(localFilePath);
+        })
+        
+        fs.unlinkSync(localFilePath)
         return response;
+
     } catch (error) {
+    console.log("Cloudinary upload error:");
+
+    if (fs.existsSync(localFilePath)) {
         fs.unlinkSync(localFilePath);
-        return null;
     }
+    return null;
+}
 }
 
 module.exports = uploadOnCloudinary;
